pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "environmental_stac_generator"
version = "0.0.1"
description = "Preprocess machine learning weather forecast netCDF prediction outputs to GeoTIFF for use with environmental-stac-dashboard"
readme = "README.md"
authors = [
  {name = "Bryn Noel Ubald", email = "bryald@bas.ac.uk"}
]
maintainers = [
  {name = "Bryn Noel Ubald", email = "bryald@bas.ac.uk"}
]
classifiers = [
]
license = {text = "MIT license"}
dependencies = [
    "deepdiff",
    "python-dateutil",
    "matplotlib",
    "numpy",
    "prompt_toolkit==3.0.36",
    "pystac",
    "python-dotenv",
    "ipython==8.18.0",
    "rich[jupyter]",
    "rioxarray",
    "shapely",
    "tqdm",
    "typer",
    "xarray[io]",
]

[project.optional-dependencies]
dev = [
    "coverage",
    "mypy",
    "pytest",
    "ruff"
]

[project.scripts]
envstacgen = "environmental_stac_generator.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"*" = ["*.*"]
